requirements.txt
requests
pandas
tqdm
rich
python-dotenv
openpyxl
